+++ tests/RouteCollectionTest.php
@@ -36,7 +36,7 @@ final class RouteCollectionTest extends TestCase
         $group = Group::create()->routes($listRoute, $viewRoute);
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage("A route with name 'my-route' already exists.");
@@ -52,7 +52,7 @@ final class RouteCollectionTest extends TestCase
         $group = Group::create()->routes($listRoute);
 
         $collector = new RouteCollector();
-        $collector->addItem($group, $viewRoute);
+        $collector->addRoute($group, $viewRoute);
 
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage("A route with name 'my-route' already exists.");
@@ -67,7 +67,7 @@ final class RouteCollectionTest extends TestCase
         $group = Group::create()->routes($listRoute);
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $this->expectException(RouteNotFoundException::class);
         $this->expectExceptionMessage('Cannot generate URI for route "wrong-name"; route not found');
@@ -85,7 +85,7 @@ final class RouteCollectionTest extends TestCase
         $group = Group::create()->routes($listRoute, $viewRoute);
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route = $routeCollection->getRoute('my-route');
@@ -104,7 +104,7 @@ final class RouteCollectionTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route = $routeCollection->getRoute('image');
@@ -141,7 +141,7 @@ final class RouteCollectionTest extends TestCase
             ->namePrefix('/api');
 
         $collector = new RouteCollector();
-        $collector->addItem($group1, $group2);
+        $collector->addRoute($group1, $group2);
 
         $routeCollection = new RouteCollection($collector);
         $routeTree = $routeCollection->getRouteTree();
@@ -175,7 +175,7 @@ final class RouteCollectionTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $routes = $routeCollection->getRoutes();
@@ -201,7 +201,7 @@ final class RouteCollectionTest extends TestCase
             ->host('https://yiiframework.com/');
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route1 = $routeCollection->getRoute('image');
@@ -232,7 +232,7 @@ final class RouteCollectionTest extends TestCase
             )->namePrefix('api');
 
         $collector = new RouteCollector();
-        $collector->addItem($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route1 = $routeCollection->getRoute('api/post/view');
@@ -270,7 +270,7 @@ final class RouteCollectionTest extends TestCase
 
         $collector = new RouteCollector();
         $collector->middleware($middleware);
-        $collector->addItem($group, $viewRoute);
+        $collector->addRoute($group, $viewRoute);
 
         $routeCollection = new RouteCollection($collector);
         $route1 = $routeCollection->getRoute('list');
@@ -322,7 +322,7 @@ final class RouteCollectionTest extends TestCase
             ->action([TestController::class, 'index'])
             ->name('main');
 
-        $collector->addItem(
+        $collector->addRoute(
             $groupWrapped ? Group::create()->routes($rawRoute) : $rawRoute
         );
 
@@ -349,7 +349,7 @@ final class RouteCollectionTest extends TestCase
         $collector = new RouteCollector();
         $collector->middleware(TestMiddleware1::class);
 
-        $collector->addItem(
+        $collector->addRoute(
             Route::get('i/{image}')->name('image')
         );