+++ tests/GroupTest.php
@@ -107,7 +107,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route = $routeCollection->getRoute('request1');
@@ -142,7 +142,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route = $routeCollection->getRoute('request1');
@@ -171,7 +171,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $route = $routeCollection->getRoute('request1');
@@ -319,7 +319,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
         $routeCollection = new RouteCollection($collector);
 
         $this->assertCount(3, $routeCollection->getRoutes());
@@ -341,7 +341,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
         $routeCollection = new RouteCollection($collector);
 
         $this->assertCount(4, $routeCollection->getRoutes());
@@ -366,7 +366,7 @@ final class GroupTest extends TestCase
             );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
         $routeCollection = new RouteCollection($collector);
 
         $this->assertCount(5, $routeCollection->getRoutes());
@@ -391,7 +391,7 @@ final class GroupTest extends TestCase
         );
 
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $this->assertCount(7, $routeCollection->getRoutes());
@@ -415,7 +415,7 @@ final class GroupTest extends TestCase
             static fn () => new Response(204)
         );
         $collector = new RouteCollector();
-        $collector->addGroup($group);
+        $collector->addRoute($group);
 
         $routeCollection = new RouteCollection($collector);
         $this->assertCount(8, $routeCollection->getRoutes());