+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Yii Router Change Log
 
-## 3.0.1 under development
+## 4.0.0 under development
 
+- Chg #207: Replace two `RouteCollectorInterface` methods `addRoute()` and `addGroup()` to single `addRoute()` (@vjik)
 - Enh #202: Add support for `psr/http-message` version `^2.0` (@vjik)
 
 ## 3.0.0 February 17, 2023